API work complete for practice

+++ controllers/api/v1/posts_api.js
@@ -1,17 +1,17 @@
 const Post = require('../../../models/post');
 const Comment = require('../../../models/comment');
 
-module.exports.index = async (req,res)=>{
+module.exports.index = async (req, res) => {
 
     let posts = await Post.find({})
-      .sort('-createdAt')
-      .populate('user')
-      .populate({
-        path: 'comments',
-        populate: {
-          path: 'user'
-        }
-      });
+        .sort('-createdAt')
+        .populate('user')
+        .populate({
+            path: 'comments',
+            populate: {
+                path: 'user'
+            }
+        });
 
 
     return res.json(200, {
@@ -26,23 +26,25 @@ module.exports.destroy = async (req, res) => {
     try {
         let post = await Post.findById(req.params.id);
 
-        // if (post.user == req.user.id) {
+        if (post.user == req.user.id) {
             post.remove();
 
             await Comment.deleteMany({ post: req.params.id });
-        
-        return res.json(200, {
+
+            return res.json(200, {
                 message: "post and associated commnets deleted succ"
             });
-        // } else {
-        //     req.flash('success', 'You can not delete this post');
+        } else {
 
-        //     return res.redirect('back');
-        // }
+            return res.json(401, {
+                message: "you cannot delete this post!"
+
+            });
+        }
     } catch (err) {
-console.log(err);
+        console.log(err);
         // req.flash('error', err);
-        return res.json(500,{
+        return res.json(500, {
             message: 'Internal server error'
         });
